fix(post-hog-event): fix renaming to event
