+++ src/core/events/WebSocketEvents.tsx
@@ -3,7 +3,8 @@ import { usePostHog } from 'posthog-js/react'
 import { PropsWithChildren, useCallback, useEffect, useRef } from 'react'
 import { useUserProfile } from 'src/core/auth'
 import { useSnackbar } from 'src/core/snackbar'
-import { PostHogEvent, apiMessages, endPoints, env } from 'src/shared/constants'
+import { apiMessages, endPoints, env } from 'src/shared/constants'
+import { PostHogEvent } from 'src/shared/post-hog'
 import { WebSocketEvent } from 'src/shared/types/server'
 import { isAuthenticated as getIsAuthenticated } from 'src/shared/utils/cookie'
 import { getAuthData } from 'src/shared/utils/localstorage'
