+++ src/pages/panel/inventory/InventoryTable.tsx
@@ -10,10 +10,11 @@ import { useEffect, useRef, useState } from 'react'
 import { useUserProfile } from 'src/core/auth'
 import { postWorkspaceInventorySearchTableQuery } from 'src/pages/panel/shared/queries'
 import { useAbsoluteNavigate } from 'src/shared/absolute-navigate'
-import { PostHogEvent, panelUI, settingsStorageKeys } from 'src/shared/constants'
+import { panelUI, settingsStorageKeys } from 'src/shared/constants'
 import { useFixQueryParser } from 'src/shared/fix-query-parser'
 import { AdvancedTableView } from 'src/shared/layouts/panel-layout'
 import { LoadingSuspenseFallback } from 'src/shared/loading'
+import { PostHogEvent } from 'src/shared/post-hog'
 import {
   PostWorkspaceInventorySearchTableResponse,
   WorkspaceInventorySearchTableColumn,
