+++ src/shared/constants/index.ts
@@ -5,4 +5,3 @@ export { endPoints } from './endPoints'
 export { env } from './env'
 export { langs } from './langs'
 export { panelUI } from './panelUI'
-export { PostHogEvent } from './postHog'
