+++ src/shared/error-boundary-fallback/ErrorBoundaryFallback.tsx
@@ -7,7 +7,8 @@ import { useEffect } from 'react'
 import { FallbackProps } from 'react-error-boundary'
 import { DiscordIcon } from 'src/assets/icons'
 import { useAbsoluteNavigate } from 'src/shared/absolute-navigate'
-import { PostHogEvent, env, panelUI } from 'src/shared/constants'
+import { env, panelUI } from 'src/shared/constants'
+import { PostHogEvent } from 'src/shared/post-hog'
 import { isAuthenticated } from 'src/shared/utils/cookie'
 import { getAuthData } from 'src/shared/utils/localstorage'
