+++ src/shared/post-hog/PostHogEvent.ts
