+++ src/shared/post-hog/index.ts
@@ -1,2 +1,3 @@
+export { PostHogEvent } from './PostHogEvent'
 export { PostHogPageView } from './PostHogPageView'
 export { PostHogProvider } from './PostHogProvider'
